+++ R/ReadJAFROCNewFormat.R
@@ -256,10 +256,14 @@ ReadJAFROCNewFormat <- function(fileName, sequentialNames)
     # this completes the move of the ratings; replace the moved ratings
     # with negative infinities
     NL[,,(K1+1):(K1+K2),] <- -Inf
-    # the following check may be superflous
+    # the following check assumes that if a case does not
+    # appear in TP sheet it must appear in FP sheet, i.e., the forced mark in 
+    # LROC paradigm; if it is not forced, then it is possible for a mark to not
+    # appear in either TP or FP sheet
     x1 <- LL_IL
     x2 <- LL
-    if (any(which(x1 != -Inf) != which(x2 == -Inf))) stop("Error in LROC file")
+    # following line should be commented if one mark is not forced
+    # if (any(which(x1 != -Inf) != which(x2 == -Inf))) stop("Error in LROC file")
     # return the LROC dataset object
     return(convert2dataset(NL, LL, LL_IL, 
                            perCase, IDs, weights,